+++ src/main/java/ar/edu/unlam/nuralign/infrastructure/dtos/LoginData.java
@@ -1,0 +1,13 @@
+package ar.edu.unlam.nuralign.infrastructure.dtos;
+
+import lombok.Builder;
+import lombok.Data;
+import lombok.ToString;
+
+@Data
+@ToString
+@Builder
+public class LoginData {
+    private String email;
+    private String password;
+}