+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # NurAlign Changelog
 
-## [3.6.1] - 2024-06-23
+## [3.6.2] - 2024-06-25
+### Changed
+- Change version of "changelog" and "pom.xml" and "application.properties".
+
+### Added
+- New features for Login Therapist on web application.
+
+
+## [3.6.1] - 2024-06-24
 ### Changed
 - Change version of "changelog" and "pom.xml" and "application.properties".
 - Fix Update MoodTracker data
 
-## [3.6.0] - 2024-06-23
+## [3.6.0] - 2024-06-24
 ### Changed
 - Change version of "changelog" and "pom.xml" and "application.properties".
 - Fix & Update TherapySession data